requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
kafka-python==2.0.2
pyspark==3.5.0
numpy>=1.26.0
pandas>=2.1.0

# Machine Learning (CPU version - faster install)
torch>=2.2.0
torchvision>=0.17.0

# Monitoring
prometheus-client==0.19.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
